requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
shap>=0.43.0
xgboost>=2.0.0
matplotlib>=3.7.0
plotly>=5.18.0
fpdf2>=2.7.0
dpkt>=1.9.0
umap-learn>=0.5.0
openpyxl>=3.1.0
